Update referral.xml
Updated the referral xml file only. Updated the ICD code section to make the code 2019 instead of 2016. The code was also changed to reflect "Malaria in Pregnancy" named "Protozoal disease complicating pregnancy, childbirth, and the puerperium" in place of "prolonged second stage of labour" in original. This is to make consistent with the Sierra Leone digital health strategy use case.